Move helper modules into helpers package
